% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschFit.R
\name{checkInput}
\alias{checkInput}
\title{Function to check user-specific Input for the right format}
\usage{
checkInput(manuelInput, resp.var, group.var)
}
\arguments{
\item{manuelInput}{manual Data input as list with response matrix and grouping vector}

\item{resp.var}{name of the binary response matrix}

\item{group.var}{name of the binary grouping vector}
}
\value{
Data ready for anchorpoint::raschFit function
}
\description{
Function to check user-specific Input for the right format
}
