% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDimDGP.R
\name{dgp_multi}
\alias{dgp_multi}
\title{Data generating process for multi dimensional rasch model (only two dimensions are supported at the moment)}
\usage{
dgp_multi(
  nobs,
  tlength,
  DIFpercent,
  Nr.dim = 2,
  Theta = NULL,
  a.vec = NULL,
  d.vec = NULL,
  DIF_mode = "intersect",
  d_distr = list(mean = 0, sd = 0.2),
  MultiNorm = NULL,
  itemtype = "dich"
)
}
\arguments{
\item{nobs}{positive interger, number of total observations (default 1000) or positive integer vector vector of length 2, number of group observations}

\item{tlength}{positive interger, number of items (default 30)}

\item{DIFpercent}{numeric array, proportion of items which have DIF per dimension (default {0,1/3})}

\item{Nr.dim}{positive interger, number of dimensions (default 2)}

\item{Theta}{matrix of the underlying ability parameters (optional)}

\item{a.vec}{numeric nobs x Nr.dim-dimensional matrix, item discrimination parameter matrix (if NA, calculated according to DIFmode & DIFpercent)}

\item{d.vec}{numeric one-dimensional vector, task difficulty matrix (if NA, calculated according to d_distr)}

\item{DIF_mode}{positive interger, number of dimensions (default 2)}

\item{d_distr}{distribution parameters normal distribution to calculate d.vec, default: mean = 0, sd = .2}

\item{MultiNorm}{list with parameters for multivariate normal which models the abilities of group 1 and group 2, respectively}

\item{itemtype}{type of items (default "dich" which corresponds to multidimensional Rasch model items)}
}
\value{
list consisting of:
\itemize{
\item binary response matrix
\item group vector (factor),
\item a.vec,
\item d.vec,
\item DIFindex,
\item Theta,
\item DIFside
}
}
\description{
Data generating process for multi dimensional rasch model (only two dimensions are supported at the moment)
}
