% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneDimDGP.R
\name{dgp_single}
\alias{dgp_single}
\title{Data generating process for one dimensional rasch model}
\usage{
dgp_single(
  nobs,
  tlength,
  DIFpercent,
  DIFpattern = "balanced",
  DIFeffect = "constant",
  DIFamount = 0.6,
  ability = TRUE,
  sigmaable = c(1, 1),
  itemref = c(-2.522, -1.902, -1.351, -1.092, -0.234, -0.317, 0.037, 0.268, -0.571,
    0.317, 0.295, 0.778, 1.514, 1.744, 1.951, -1.152, -0.526, 1.104, 0.961, 1.314,
    -2.198, -1.621, -0.761, -1.179, -0.61, -0.291, 0.067, 0.706, -2.713, 0.213, 0.116,
    0.273, 0.84, 0.745, 1.485, -1.208, 0.189, 0.345, 0.962, 1.592)
)
}
\arguments{
\item{nobs}{number of observations in each group (here equal)}

\item{tlength}{number of items the test contains}

\item{DIFpercent}{rel. frequency of DIF items in the test}

\item{DIFpattern}{"balanced": DIF balanced over groups "favorref","favorfoc": all DIF items favor one group}

\item{DIFeffect}{data generating process for DIF effect: "normal" -> Wang(2012), "uniform" -> alternative}

\item{DIFamount}{magnitude of DIF}

\item{ability}{ability}

\item{sigmaable}{positive numeric vector of length two (two groups), default c(1,1), standard deviation for person parameter matrix (abilities)}

\item{itemref}{numeric vector of length tlength (if shorter, then sampling with replacement is used), items difficulty parameters according to Hartmann}
}
\value{
list containing response matrix, group assignments, DIFindex, DIFside, itemref, itemfoc
}
\description{
Data generating process for one dimensional rasch model
}
