% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniDimDGP.R
\name{dgp_uni}
\alias{dgp_uni}
\title{Data generating process for unidimensional rasch model}
\usage{
dgp_uni(
  nobs,
  tlength,
  DIFpercent,
  DIFpattern = "balanced",
  DIFeffect = "constant",
  DIFamount = 0.6,
  ability = TRUE,
  sigmaable = c(1, 1),
  itemref = c(-2.522, -1.902, -1.351, -1.092, -0.234, -0.317, 0.037, 0.268, -0.571,
    0.317, 0.295, 0.778, 1.514, 1.744, 1.951, -1.152, -0.526, 1.104, 0.961, 1.314,
    -2.198, -1.621, -0.761, -1.179, -0.61, -0.291, 0.067, 0.706, -2.713, 0.213, 0.116,
    0.273, 0.84, 0.745, 1.485, -1.208, 0.189, 0.345, 0.962, 1.592)
)
}
\arguments{
\item{nobs}{number of observations per group}

\item{tlength}{interger > 0, test length (number of items)}

\item{DIFpercent}{percentage of DIF items in the test}

\item{DIFpattern}{"balanced": DIF balanced over groups "favorref","favorfoc": all DIF items favor one group}

\item{DIFeffect}{data generating process for DIF effect:
\itemize{
\item normal:    item parameter differences are drawn at random from a normal distribution with mean DIFamount and sd = 0.1, like in Wang et al. (2012)
\item uniform:   item parameter differences are drawn at random from the vector \code{[DIFamount-0.4,DIFamount-0.2,DIFamount,DIFamount+0.2,DIFamount+0.4]}
\item constant:  item parameter differences are defined as \code{DIFamount} for all items
}}

\item{DIFamount}{magnitude of DIF}

\item{ability}{should the groups differ in mean ability? (default is TRUE)}

\item{sigmaable}{positive numeric vector of length two, standard deviations for person parameter distributions in the two groups (default is c(1,1))}

\item{itemref}{numeric vector of length tlength (if shorter, then sampling with replacement is used), item difficulty parameter for reference group like in Wang et al. (2012)}
}
\value{
list containing:
\itemize{
\item dat: binary response matrix
\item DIFindex: indicating which items were generated with DIF
\item DIFside: which group is favored per item (-1 focal, 1 reference) default: focal group is favored for all items
\item itemref: item difficulty parameter for reference group
\item itemfoc: item difficulty parameter for focal group
\item groups: group vector (factor),
}
}
\description{
Data generating process for unidimensional rasch model
}
\examples{
# For examples, see ?getData.
}
\references{
\itemize{
\item Wang WC, Shih CL, Sun GW (2012). “The DIF-Free-Then-DIF Strategy for the Assessmentof Differential Item Functioning.”Educational and Psychological Measurement,72(4), 687–708
}
}
