% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diftests.R
\name{diftests}
\alias{diftests}
\title{Extend diftests function of psychotools to include offset}
\usage{
diftests(obj1, obj2, anchor_items, adjust, offset = 0)
}
\arguments{
\item{obj1}{\itemize{
\item rasch model object 1
}}

\item{obj2}{\itemize{
\item rasch model object 2
}}

\item{anchor_items}{anchor items}

\item{adjust}{p-value adjustment, (default: "none")}

\item{offset}{offset}
}
\value{
list containing test results, item paramters and covariance
}
\description{
Extend diftests function of psychotools to include offset
}
