% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGrid.R
\name{generateGrid}
\alias{generateGrid}
\title{This function generates the grid values.}
\usage{
generateGrid(
  beta1,
  beta2,
  grid_method = c("symmetric", "sparse"),
  j.length = 1000
)
}
\arguments{
\item{beta1}{Coefficients from first Rasch model fit}

\item{beta2}{Coefficients from second Rasch model fit}

\item{grid_method}{a string, specifying the grid method that is used to generated the shifts for evaluation ("symmetric" and/or "sparse", abbreviations are accepted)}

\item{j.length}{positive integer, the granularity of the symmetric grid (default: 1000)}
}
\value{
A list with the selected grid methods each having two components:
\enumerate{
\item A vector 'c_grid' with the grid values
\item A list of matrices 'betas_grid' for each grid value (length = j.length)
each matrix has three columns: coefficient from group 1 (beta1), shifted coefficient from group 2 (newbeta2) and distance between them (beta1-newbeta2).
The rows correspond to the items.
}
}
\description{
This function generates the grid values.
}
