% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDimDGP.R
\name{get_covmat}
\alias{get_covmat}
\title{Function that gives back a covariance matrix for n dimesnions}
\usage{
get_covmat(Nr.dim, variances = 0.25, covariances = 0.125)
}
\arguments{
\item{Nr.dim}{integer - the number of dimensions}

\item{variances}{numeric, positive, <= 1, (same for all dimensions) or Nr.dim-dimensional vector - variance of each dimension}

\item{covariances}{numeric, positive, <= 1, (same for all dimensions) or choose(Nr.dim,2)-dimensional vector - covariances between dimensions}
}
\value{
covariance matrix of dimension Nr.dim x Nr.dim
}
\description{
Function that gives back a covariance matrix for n dimesnions
}
