% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{get_results}
\alias{get_results}
\title{Function to evaluate criterion values and obtain test results for a given grid and method}
\usage{
get_results(grid, shift, getTestResults, rm, metric)
}
\arguments{
\item{grid}{The grid values: output of the "generateGrid.R" function}

\item{shift}{the desired shift}

\item{getTestResults}{logic, whether test should be applied}

\item{rm}{list containing the two Rasch Model corresponding two group 0 and 1}

\item{metric}{criterion to evaluate as a function}
}
\value{
a list containing the criterion evaluated at grid points and the result of the wald test
}
\description{
Function to evaluate criterion values and obtain test results for a given grid and method
}
