% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{getCriterionRes}
\alias{getCriterionRes}
\title{criterion function}
\usage{
getCriterionRes(
  rm1,
  rm2,
  select = c("Gini Index", "CLF Criterion"),
  grid = c("symmetric", "sparse"),
  shift = NULL
)
}
\arguments{
\item{rm1}{Fitted Rasch Model object corresponding to the first group. Object is of class "raschmodel", produced by function \code{raschmodel} of the package \pkg{psychotools}.}

\item{rm2}{Fitted Rasch Model object corresponding to the second group. Object is of class "raschmodel", produced by function \code{raschmodel} of the package \pkg{psychotools}.}

\item{select}{criterion: Gini Index or CLF Criterion}

\item{grid}{grid method: symmetric or sparse}

\item{shift}{desired shift. if NULL, then the criterion maximizing is used. Can also be numeric to get desired shift. Caution: must be within grid!}
}
\value{
a list "resultlist" which contains: grid values, criterion values, rm object, coefficients and information about generating mechanism
}
\description{
criterion function
}
