% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Simulate data from one dimensional or multidimensional DGP}
\usage{
getData(nobs, tlength, DIFpercent, type = c("single", "multi"), ...)
}
\arguments{
\item{nobs}{positive integer, number of observations}

\item{tlength}{positive integer, number of items}

\item{DIFpercent}{non-negative double <= 1, DIF percentage}

\item{type}{string specifying which DGP to use ("single" or "multi")}

\item{...}{additional arguments:
\itemize{
\item For single-dimensional DGP, check ?anchorpoint::dgp_single
\item For multi-dimensional DGP, check ?anchorpoint::dgp_multi
}}
}
\value{
a list containing
\itemize{
\item DGP: simulated data
\item RM: Rasch Model objects
}
}
\description{
Simulate data from one dimensional or multidimensional DGP
}
