% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemDiscrimination.R
\name{getItemDiscrimination}
\alias{getItemDiscrimination}
\title{Function to create a item discrimination parameter matrix}
\usage{
getItemDiscrimination(
  dimensions,
  DIFpercent,
  tlength,
  DIF_mode = c("intersect", "disjoint")
)
}
\arguments{
\item{dimensions}{integer specifying the number of dimensions used in dgp (currently only 2 are allowed)}

\item{DIFpercent}{numeric, quantifying DIF percentage}

\item{tlength}{positive integer specifying the number of items}

\item{DIF_mode}{character vector specifying the mode how to create the matrix:
\itemize{
\item "intersect":       all items load on the first, length*DIFpercent items also on the second
\item "disjoint":        ceiling(tlength*DIFpercent) items load on the first, the rest on the second, where ceiling rounds the number up to the next integer
}}
}
\value{
A binary item discrimination parameter matrix of dimension: tlength x dimensions
}
\description{
Function to create a item discrimination parameter matrix
}
