% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWald.R
\name{getWald}
\alias{getWald}
\title{Function which executes Wald test for given rm object and shift (with "min_dist" setting)}
\usage{
getWald(rm, shift)
}
\arguments{
\item{rm}{A list containing the two Rasch Model objects of group 0 and group 1}

\item{shift}{Shift in item parameters for the second group}
}
\value{
A list containing the output of the function Wald_test():
\itemize{
\item p: results from the test (p-values)
\item vcov: the covariance matrices of the fit (from diftests function)
}
}
\description{
Function which executes Wald test for given rm object and shift (with "min_dist" setting)
}
