% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{graphicalTest}
\alias{graphicalTest}
\title{Function to produce graphical test plot}
\usage{
graphicalTest(
  object,
  shift = NULL,
  highlight = NULL,
  alpha = 0.05,
  testColors = list(`not significant` = "darkgreen", significant = "red3", `anchor\\n  
                                             item` = "black"),
  TestResults = NULL,
  ask = TRUE,
  ...
)
}
\arguments{
\item{object}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{shift}{numeric, shift which is applied, default NULL (global optimum)}

\item{highlight}{positive integer, items to be highlighted (invalid items are excluded).}

\item{alpha}{significance level for DIF test}

\item{testColors}{list with colors for the items:
\itemize{
\item "not significant" = "darkgreen"
\item "significant" = "red3"
\item "anchor item" = "black"
}}

\item{TestResults}{waldtest object from anchorpoint::getWald. If NULL, then they are computed within the function. Default: NULL.}

\item{ask}{logical, ask for next plot. Default = TRUE}

\item{...}{further arguments for plot() like lty, cex.axis, cex.main, cex.lab etc.
credit: some code is adapted from the function plotGOF of the R package eRm.}
}
\description{
Function to produce graphical test plot
}
