% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{plot.anchorpoint}
\alias{plot.anchorpoint}
\title{plot function
#' hand over location_picker = TRUE, to identify specific points in the plot
to terminate the function, press any mouse button other than the first (X11 device) or press ESC key (quartz)
see ?identify for help}
\usage{
\method{plot}{anchorpoint}(x, ask = T, location_picker = FALSE, ...)
}
\arguments{
\item{x}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{ask}{logical, ask for next plot. Default = TRUE}

\item{location_picker}{logical, use location picker. Default FALSE.}

\item{...}{additional parameters for plot function as for standard plot function (e.g. col)}
}
\description{
plot function
#' hand over location_picker = TRUE, to identify specific points in the plot
to terminate the function, press any mouse button other than the first (X11 device) or press ESC key (quartz)
see ?identify for help
}
