% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{plotCriterion}
\alias{plotCriterion}
\title{Function to produce criterion plot}
\usage{
plotCriterion(
  object,
  names,
  location_picker = FALSE,
  lty = 1,
  col = 1,
  cex.axis = 1,
  cex.lab = 1,
  cex.main = 1,
  cex = 1
)
}
\arguments{
\item{object}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{names}{list, with criterion and grid: names of the methods used.}

\item{location_picker}{use location picker}

\item{lty}{line type}

\item{col}{color}

\item{cex.axis}{cex.axis}

\item{cex.lab}{cex.lab}

\item{cex.main}{cex.main}

\item{cex}{cex}
}
\value{
selected points with additional information
}
\description{
Function to produce criterion plot
}
