% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{print.anchorpoint}
\alias{print.anchorpoint}
\title{Print function}
\usage{
\method{print}{anchorpoint}(x, ...)
}
\arguments{
\item{x}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{...}{further arguments passed to or from other methods (e.g. digits for rounding).}
}
\description{
Print function
}
