% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{print.plot.anchorpoint}
\alias{print.plot.anchorpoint}
\title{print function for plot.anchorpoint}
\usage{
\method{print}{plot.anchorpoint}(x, ...)
}
\arguments{
\item{x}{plot.anchorpoint object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print function for plot.anchorpoint
}
