% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{print.WaldtestpV}
\alias{print.WaldtestpV}
\title{Print function for WaldtestpV object}
\usage{
\method{print}{WaldtestpV}(x, ...)
}
\arguments{
\item{x}{Wald test object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for WaldtestpV object
}
