% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{shiftPlot}
\alias{shiftPlot}
\title{Function to produce shift Plot}
\usage{
shiftPlot(
  object,
  shift = NULL,
  testColors = list(`not significant` = "darkgreen", significant = "red3", `anchor
    item` = "black"),
  testPCH = list(`not significant` = 21, significant = 22, `anchor item` = 23),
  addLegend = TRUE,
  highlight = NULL,
  digits = 3,
  cex.legend = 0.5,
  TestResults = NULL,
  ask = TRUE,
  ...
)
}
\arguments{
\item{object}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{shift}{numeric, shift which is applied}

\item{testColors}{list with colors for the items:
\itemize{
\item "not significant" = "darkgreen"
\item "significant" = "red3"
\item "anchor item" = "black"
}}

\item{testPCH}{list with pch for the items (for color blind people):
\itemize{
\item "not significant" = 21
\item "significant" = 22
\item "anchor item" = 23
}}

\item{addLegend}{logic, add a legend to the plot, default: False}

\item{highlight}{positive integer, items to be highlighted (invalid items are excluded).}

\item{digits}{positive integer, controls rounding of the shift in title}

\item{cex.legend}{numeric,  controls size of legend}

\item{TestResults}{waldtest object from anchorpoint::getWald. If NULL, then they are computed within the function. Default: NULL.}

\item{ask}{logical, ask for next plot. Default = TRUE}

\item{...}{additional graphics arguments}
}
\description{
Function to produce shift Plot
}
