% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{summary.anchorpoint}
\alias{summary.anchorpoint}
\title{summary function}
\usage{
\method{summary}{anchorpoint}(object, ...)
}
\arguments{
\item{object}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
summary function
}
