% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald_test.R
\name{Wald_test}
\alias{Wald_test}
\title{Function to create a item discrimination parameter matrix}
\usage{
Wald_test(
  rm1,
  rm2,
  c_shift,
  alias_method = c("constant4_MPT", "quasi_var", "min_dist"),
  alias_anchor_items = NULL,
  adjust = "none"
)
}
\arguments{
\item{rm1}{Fitted Rasch Model object corresponding to the first group. Object is of class "raschmodel", produced by function \code{RaschModel.fit} of the package \pkg{psychotools}.}

\item{rm2}{Fitted Rasch Model object corresponding to the first group. Object is of class "raschmodel", produced by function \code{RaschModel.fit} of the package \pkg{psychotools}.}

\item{c_shift}{The shift of the second group}

\item{alias_method}{character vector specifying the aliasing method. One of "constant4_MPT", "quasi_var", "min_dist".}

\item{alias_anchor_items}{integer in 1,...,#items. Default: NULL, will be chosen according to alias_method}

\item{adjust}{p-value adjustment (multiple testing correction), Default: "none"}
}
\value{
list containing
\itemize{
\item p: results from the test (p-values)
\item vcov: the covariance matrices of the fit (from diftests function)
}
}
\description{
Function to create a item discrimination parameter matrix
}
