% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{waldtestpV}
\alias{waldtestpV}
\title{Function to get waldtest p-value results}
\usage{
waldtestpV(object, shift = NULL, ...)
}
\arguments{
\item{object}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{shift}{shift, default NULL (global optimum), else numeric}

\item{...}{further arguments for signif(x,...) (digits)}
}
\description{
Function to get waldtest p-value results
}
