% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{anchorpoint}
\alias{anchorpoint}
\title{Function to produce anchorpoint objects}
\usage{
anchorpoint(
  rm1,
  rm2,
  select = c("CLF Criterion", "Gini Index"),
  grid = c("symmetric", "sparse")
)
}
\arguments{
\item{rm1}{Fitted Rasch Model object corresponding to the first group. Object is of class "raschmodel", produced by function \code{RaschModel.fit} of the package \pkg{psychotools}.}

\item{rm2}{Fitted Rasch Model object corresponding to the second group. Object is of class "raschmodel", produced by function \code{RaschModel.fit} of the package \pkg{psychotools}.}

\item{select}{a string, specifying the criterion that is evaluated ("CLF Criterion" or "Gini Index", abbreviations are accepted)}

\item{grid}{a string, specifying the grid method that is used to generated the shifts for evaluation ("symmetric" or "sparse", abbreviations are accepted)}
}
\value{
an anchorpoint object containing:
\itemize{
\item list with global optima
\item list with all results (grids and criterion values)
\item string with used criteria
\item string with used grids
\item list with the rm object
}
}
\description{
Function to produce anchorpoint objects
}
