% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clfCriterion.R
\name{clfCriterion}
\alias{clfCriterion}
\title{calculates the CLF criterion}
\usage{
clfCriterion(dist, eps = 1e-04)
}
\arguments{
\item{dist}{distance vector}

\item{eps}{shift for numerical stability (see Muthen paper)}
}
\value{
criterion value
}
\description{
calculates the CLF criterion
}
