% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDimDGP.R
\name{dgp_multi}
\alias{dgp_multi}
\title{Data generating process for multidimensional Rasch model (only two dimensions are supported at the moment)}
\usage{
dgp_multi(
  nobs,
  tlength,
  DIFpercent,
  Nr.dim = 2,
  Theta = NULL,
  discriminations = NULL,
  difficulties = NULL,
  DIF_mode = "intersect",
  d_distr = list(mean = 0, sd = 0.2),
  MultiNorm = NULL,
  itemtype = "dich"
)
}
\arguments{
\item{nobs}{positive integer, number of total observations (default 1000) or positive integer vector vector of length 2, number of observations per group}

\item{tlength}{interger > 0, test length (number of items)}

\item{DIFpercent}{percentage of DIF items in the test}

\item{Nr.dim}{positive integer, number of dimensions (default 2)}

\item{Theta}{matrix of the underlying ability parameters (optional)}

\item{discriminations}{binary matrix of size tlength x Nr.dim, item discrimination parameter matrix}

\item{difficulties}{numeric vector of length tlength, item difficulty vector}

\item{DIF_mode}{string, mode how DIF items are created, default: intersect.}

\item{d_distr}{d_distr: parameters for normal distribution to generate item difficulties difficulties, default: mean = 0, sd = .2}

\item{MultiNorm}{list with parameters for multivariate normal distribution to generate the abilities of the test takers in group 1 and group 2, respectively.}

\item{itemtype}{type of items (default "dich" which corresponds to multidimensional Rasch model items)}
}
\value{
list consisting of:
\itemize{
\item dat: binary response matrix
\item groups: group vector (factor),
\item discriminations: binary matrix containing the item discrimination parameter matrix,
\item difficulties: item difficulty vector,
\item DIFindex: indicating which items were generated with DIF,
\item Theta,
\item DIFside: which group is favored, default focal group is favored
}
}
\description{
Data generating process for multidimensional Rasch model (only two dimensions are supported at the moment)
}
\references{
\itemize{
\item Data is generated using the function simdata from \pkg{mirt} (Version: 1.32.1).
}
}
