% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Simulate data from unidimensional or multidimensional DGP}
\usage{
getData(nobs, tlength, DIFpercent, type = c("uni", "multi"), ...)
}
\arguments{
\item{nobs}{positive integer, number of observations}

\item{tlength}{interger > 0, test length (number of items)}

\item{DIFpercent}{percentage of DIF items in the test}

\item{type}{string specifying which DGP to use ("uni" or "multi")}

\item{...}{additional arguments:
\itemize{
\item For unidimensional DGP, check ?anchorpoint:::dgp_uni
\item For multidimensional DGP, check ?anchorpoint:::dgp_multi
}}
}
\value{
a list containing:
\itemize{
\item DGP: simulated data
\item RM: Rasch Model objects for the two groups of test takers
}
}
\description{
Simulate data from unidimensional or multidimensional Rasch model and two groups of test takers.
}
\examples{
# The number of observations
nobs = 20
# The number of items
tlength = 10
# The percentage of items to experience differential item functioning
DIFpercent = 0

# Create data from a unidimensional DGP:
getData(nobs, tlength, DIFpercent, type = "uni")
}
