% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{graphicalTest}
\alias{graphicalTest}
\title{Function to produce graphical test plot}
\usage{
graphicalTest(
  object,
  shift = NULL,
  highlight = NULL,
  alpha = 0.05,
  testColors = list(`not significant` = "darkgreen", significant = "red3", `anchor
    item` = "black"),
  TestResults = NULL,
  ask = TRUE,
  ...
)
}
\arguments{
\item{object}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{shift}{shift in item parameters for the second group, default NULL (for global optimum), else numeric (for user-defined shift)}

\item{highlight}{positive integer(s), numbers of the items to be highlighted}

\item{alpha}{significance level for DIF test}

\item{testColors}{list with colors for the items:
\itemize{
\item "not significant" = "darkgreen"
\item "significant" = "red3"
\item "anchor item" = "black"
}}

\item{TestResults}{Waldtest object from anchorpoint::getWald. If NULL, then they are computed within the function. Default: NULL.}

\item{ask}{logical, ask for next plot. Default = TRUE}

\item{...}{further arguments for plot() like lty, cex.axis, cex.main, cex.lab etc.}
}
\description{
Function to produce graphical test plot
}
\examples{
# Load the SPISA data set (general knowledge quiz - more information at ?SPISA)
library("psychotree")
data("SPISA")

# Fit the Rasch Models for the two groups females and males
fit <- raschFit(SPISA, resp.mat.name='spisa', group.name='gender')

# Rasch Model fit for the first and second group
rm1 <- fit$rm1
rm2 <- fit$rm2

# Fit an Anchorpoint object
ap_object <- anchorpoint(rm1,rm2,select = "Gini Index", grid = "sparse")

# Use the Anchorpoint object to get the graphical test
graphicalTest(ap_object)
}
\references{
Credit: Part of the code is adapted from the function \code{plotGOF} of the package \pkg{eRm} (Version: 1.32.1):
\itemize{
\item Mair P, Hatzinger R. Extended Rasch modeling: The eRm package for the application of IRT models in R. Journal of Statistical Software. 2007;20 (9) :1-20.
}
}
