% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{print.summary.anchorpoint}
\alias{print.summary.anchorpoint}
\title{print function for summary.anchorpoint}
\usage{
\method{print}{summary.anchorpoint}(x, ...)
}
\arguments{
\item{x}{summary.anchorpoint object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print function for summary.anchorpoint
}
