% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{print.waldtestpV}
\alias{print.waldtestpV}
\title{Print function for waldtestpV object}
\usage{
\method{print}{waldtestpV}(x, ...)
}
\arguments{
\item{x}{waldtest object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for waldtestpV object
}
