% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{shiftPlot}
\alias{shiftPlot}
\title{Function to produce a shift plot}
\usage{
shiftPlot(
  object,
  shift = NULL,
  testColors = list(`not significant` = "darkgreen", significant = "red3", anchoritem =
    "black"),
  testPCH = list(`not significant` = 21, significant = 22, anchoritem = 23),
  addLegend = TRUE,
  highlight = NULL,
  digits = 3,
  cex.legend = 0.75,
  TestResults = NULL,
  ask = TRUE,
  ...
)
}
\arguments{
\item{object}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{shift}{shift in item parameters for the second group, default NULL (for global optimum), else numeric (for user-defined shift)}

\item{testColors}{list with colors for the items:
\itemize{
\item "not significant" = "darkgreen"
\item "significant" = "red3"
\item "anchoritem" = "black"
}}

\item{testPCH}{list with pch for the items (for color blind people):
\itemize{
\item "not significant" = 21
\item "significant" = 22
\item "anchoritem" = 23
}}

\item{addLegend}{logic, add a legend to the plot, default: False}

\item{highlight}{positive integer(s), numbers of the items to be highlighted}

\item{digits}{positive integer, controls rounding of the shift in title}

\item{cex.legend}{numeric,  controls size of legend}

\item{TestResults}{Waldtest object from anchorpoint::getWald. If NULL, then they are computed within the function. Default: NULL.}

\item{ask}{logical, ask for next plot. Default = TRUE}

\item{...}{additional graphics arguments}
}
\description{
Function to produce a shift plot
}
\examples{
# Load the SPISA data set (general knowledge quiz - more information at ?SPISA)
library("psychotree")
data("SPISA")

# Fit the Rasch Models for the two groups females and males
fit <- raschFit(SPISA, resp.mat.name='spisa', group.name='gender')

# Rasch Model fit for the first and second group
rm1 <- fit$rm1
rm2 <- fit$rm2

# Fit an Anchorpoint object
ap_object <- anchorpoint(rm1,rm2,select = "Gini Index", grid = "sparse")

# Use the Anchorpoint object to get the shift plot
shiftPlot(ap_object)
}
