% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpoint.R
\name{WaldtestpV}
\alias{WaldtestpV}
\title{Function to get Wald test p-value results}
\usage{
WaldtestpV(object, shift = NULL, ...)
}
\arguments{
\item{object}{anchorpoint object as produced by the function \code{anchorpoint}}

\item{shift}{shift in item parameters for the second group, default NULL (for global optimum), else numeric (for user-defined shift)}

\item{...}{further arguments for signif(x,...) (digits)}
}
\description{
Function to get Wald test p-value results
}
\examples{
# Load the SPISA data set (general knowledge quiz - more information at ?psychotree::SPISA)
library("psychotree")
data("SPISA")
# Fit the Rasch Models for the two groups females and males
fit <- raschFit(SPISA, resp.mat.name='spisa', group.name='gender')
# Rasch Model fit for the first group
rm1 <- fit$rm1
# Rasch Model fit for the second group
rm2 <- fit$rm2
# Fit an Anchorpoint object
ap_object <- anchorpoint(rm1,rm2,select = "Gini Index", grid = "sparse")
# Obtain the Wald test p-values
WaldtestpV(ap_object)
}
